+++ docs/source/changelog.rst
@@ -93,6 +93,10 @@ UNRELEASED
 
 * ``waitForAllNodesToFinish()`` method in ``DriverDSLExposedInterface`` has instead become a parameter on driver creation.
 
+* Enterprise Corda only: Compatibility with SQL Server 2017 and SQL Azure databases.
+
+* Enterprise Corda only: node configuration property ``database.schema`` and documented existing database properties.
+
 .. _changelog_v1:
 
 Release 1.0